requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
Faker==37.12.0
filetype==1.2.0
idna==3.11
imageio==2.37.0
iniconfig==2.1.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.3
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251219
pillow_heif==1.1.0
pillow==11.3.0
pluggy==1.6.0
pyclip==0.7.0
Pygments==2.19.2
PyInputPlus==0.2.12
pypdf==6.1.3
PySimpleValidate==0.2.12
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.2
Send2Trash==1.8.3
setuptools==80.7.1
six==1.17.0
stdiomask==0.0.6
tabulate==0.9.0
tifffile==2025.9.20
tqdm==4.67.1
types-pytz==2025.2.0.20251108
tzdata==2025.2
urllib3==2.6.2
